ktv-multilayer-api/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0
loguru==0.7.2
earthengine-api==0.1.379
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-cloud-storage==2.10.0
geojson==3.1.0
shapely==2.0.2
geopandas==0.14.1
rasterio==1.3.9
fiona==1.9.5
pyproj==3.6.1
numpy==1.24.4
pandas==2.1.4